+++ changelog.md
@@ -17,6 +17,7 @@
 - magnets now break from damagetypes in the tag `ultracraft:break_magnet`
 - Vent Covers can now be oriented vertically
 - Added Recipe for Vent Covers
+- Cerberus is now guaranteed to drop a golden Apple if a Cerberus Ball was parried back at it
 ## Settings & Gamerules
 ## Commands
 - `/ultracraft progression` subcommands now support multiple targets